terms page styling
