+++ db/seeds.rb
@@ -24,7 +24,7 @@ if Rails.env.development?
 end
 
 # temporary terms of use
-Terms.create(terms: "Lorem ipsum dolor sit amet, consectetur adipisicing elit. Eligendi, inventore delectus, laborum excepturi repudiandae, ex quae eveniet, modi culpa sequi distinctio perspiciatis voluptate sint molestiae! Iure suscipit, vitae eaque ut.", type: :tou)
+Terms.create(terms: "Lorem ipsum dolor sit amet, consectetur adipisicing elit. Eligendi, inventore delectus, laborum excepturi repudiandae, ex quae eveniet, modi culpa sequi distinctio perspiciatis voluptate sint molestiae! Iure suscipit, vitae eaque ut. Lorem ipsum dolor sit amet, consectetur adipisicing elit. Illo quibusdam eum et quis ipsum consectetur iusto eligendi labore ab quaerat in tempora fuga eaque debitis, delectus incidunt, minus alias amet. \n Lorem ipsum dolor sit amet, consectetur adipisicing elit. Ex ducimus quis voluptatem sed pariatur optio dignissimos. Eum vero quibusdam ducimus cumque earum dicta accusantium a libero aut debitis, facere repellendus! \n Lorem ipsum dolor sit amet, consectetur adipisicing elit. Doloremque provident iusto suscipit sunt incidunt explicabo laudantium culpa dolore quo totam eligendi, ea unde tenetur quidem, illo dolor neque, aliquam commodi.", type: :tou)
 Terms.create(terms: "Lorem ipsum dolor sit amet, consectetur adipisicing elit. Doloribus eius omnis expedita, eveniet id ut molestiae deserunt, libero modi magni adipisci sed ea non ipsam repellendus voluptas quia nihil labore!", type: :privacy)
 
 ############################
